:recycle: Keep trigger return SQL in MSSQL.
Move SQL Server's trigger output table, type validation, and output ordering into its query generator. Remove the single-dialect capability from core while retaining generic return field metadata.
