+++ packages/core/src/abstract-dialect/dialect.ts
@@ -229,7 +229,6 @@ export type DialectSupports = {
     unquoted: boolean;
     quoted: boolean;
   };
-  tmpTableTrigger: boolean;
   indexHints: boolean;
   tableHints: boolean;
   searchPath: boolean;
@@ -478,7 +477,6 @@ export abstract class AbstractDialect<
     },
     REGEXP: false,
     IREGEXP: false,
-    tmpTableTrigger: false,
     indexHints: false,
     tableHints: false,
     searchPath: false,
