+++ packages/mssql/src/dialect.ts
@@ -66,7 +66,6 @@ export class MsSqlDialect extends AbstractDialect<MsSqlDialectOptions, MsSqlConn
     inserts: {
       ignoreDuplicates: false,
     },
-    tmpTableTrigger: true,
     dataTypes: {
       JSON: true,
       // TODO: https://learn.microsoft.com/en-us/sql/t-sql/spatial-geography/spatial-types-geography?view=sql-server-ver16
