+++ packages/mssql/src/query-generator.js
@@ -218,6 +218,40 @@ export class MsSqlQueryGenerator extends MsSqlQueryGeneratorTypeScript {
     ]);
   }
 
+  generateReturnValues(modelAttributes, options) {
+    const returnValues = super.generateReturnValues(modelAttributes, options);
+
+    if (!options.hasTrigger) {
+      return returnValues;
+    }
+
+    if (returnValues.returnTypes.includes(undefined)) {
+      throw new Error(
+        'MSSQL bulk inserts with triggers require returning column names that match model attributes.',
+      );
+    }
+
+    let orderColumn = '__sequelize_tmp_order';
+    while (returnValues.returnFields.includes(this.quoteIdentifier(orderColumn))) {
+      orderColumn = `_${orderColumn}`;
+    }
+
+    const quotedOrderColumn = this.quoteIdentifier(orderColumn);
+    const tmpColumns = returnValues.returnFields.map((field, i) => {
+      return `${field} ${attributeTypeToSql(returnValues.returnTypes[i], {
+        dialect: this.dialect,
+      })}`;
+    });
+    tmpColumns.push(`${quotedOrderColumn} BIGINT IDENTITY(1,1)`);
+
+    return {
+      ...returnValues,
+      outputFragment: `${returnValues.outputFragment} INTO @tmp`,
+      returningFragment: `; SELECT ${returnValues.returnFields.join(', ')} FROM @tmp ORDER BY ${quotedOrderColumn}`,
+      tmpTable: `DECLARE @tmp TABLE (${tmpColumns.join(',')}); `,
+    };
+  }
+
   bulkInsertQuery(tableName, attrValueHashes, options, attributes) {
     const quotedTable = this.quoteTable(tableName);
     options ||= {};
@@ -237,7 +271,7 @@ export class MsSqlQueryGenerator extends MsSqlQueryGeneratorTypeScript {
 
       outputFragment = returnValues.outputFragment;
       returningFragment = returnValues.returningFragment;
-      tmpTable = returnValues.tmpTable;
+      tmpTable = returnValues.tmpTable ?? '';
     }
 
     const emptyQuery = `INSERT INTO ${quotedTable}${outputFragment} DEFAULT VALUES`;
